Implement high-performance identifiable key scanning in managed code
This change removes the interop with the rust API from `IdentifiableScan` and implements similar high-performance scanning techniques in managed code. This reduces the size of the NuGet package tremendously, from ~7 MB to ~200 KB, and adds support for high-performance scanning on non x86/x64 CPUs and non-Windows operating systems.

Performance is generally better than before for managed callers of `IdentifiableScan` with significant gains in most cases. See benchmarks below.

Code generation is used to avoid having to keep the high-performance scanning code in sync when patterns are changed. See also extensive code comments on the approach.

For now, the streaming capability of `IdentifiableScan` has been removed, but this can be re-added as the managed algorithm is actually streaming friendly by only ever regex-matching against short sub-ranges of input. This was held back for lack of time and discovery that our only partner that calls this API does not use the streaming.

Finally, the `StyleCop` analyzer is removed because it broke on newer C# constructs. This will be replaced with a `dotnet format` check like the one in `microsoft/cask` in a follow-up change.